Update UI components to use integer dimensions and padding
Refactored `ButtonComponent` and `VerticalListContainer` to replace `float` dimensions with `int` for better precision and consistency. Updated related methods and adjusted calculations to account for integer-based dimensions and padding. This simplifies alignment and enhances maintainability.

+++ include/ui/components/containers/vertical_list_container.h
@@ -5,29 +5,29 @@ namespace artifact
 {
     class VerticalListContainer : public ContainerBase
     {
-        float x;
-        float y;
-        float width;
-        float height;
-        float gap;
-        float padding_top;
-        float padding_bottom;
-        float padding_left;
-        float padding_right;
+        int x;
+        int y;
+        int width;
+        int height;
+        int gap;
+        int padding_top;
+        int padding_bottom;
+        int padding_left;
+        int padding_right;
         Color background_color;
 
     public:
-        VerticalListContainer(const char *identifier, float x, float y, float width, float height, float gap = 0.0f, float padding = 0.0f, Color background_color = BLANK);
+        VerticalListContainer(const char *identifier, int x, int y, int width, int height, int gap = 0, int padding = 0, Color background_color = BLANK);
 
         void draw() override;
         void update(int mouse_x, int mouse_y) override;
 
         // Setters
-        void set_gap(const float gap) { this->gap = gap; }
-        void set_padding(float padding);
-        void set_padding(float top, float bottom, float left, float right);
+        void set_gap(const int gap) { this->gap = gap; }
+        void set_padding(int padding);
+        void set_padding(int top, int bottom, int left, int right);
         void set_background_color(const Color color) { this->background_color = color; }
-        void set_position(float x, float y);
-        void set_size(float width, float height);
+        void set_position(int x, int y);
+        void set_size(int width, int height);
     };
 } // namespace artifact
